refacuring and added new component DraggableItem
