+++ frontend/form-builder/src/components/Toolbar/DraggableItem/DraggableItem.module.css
@@ -1,0 +1,3 @@
+.toolbarIcon {
+  margin-bottom: var(--fds-spacing-4);
+}
