+++ frontend/form-builder/src/components/Toolbar/DraggableItem/index.ts
@@ -1,0 +1,1 @@
+export { DraggableItem } from "./DraggableItem";
