+++ frontend/form-builder/src/components/Toolbar/ToolBarItems/ToolBarItems.tsx
@@ -1,6 +1,4 @@
 import React from "react";
-import { useDrag } from "react-dnd";
-import { DraggableItemsType } from "../../../types/dndTypes";
 import {
   BulletListIcon,
   TasklistSaveIcon,
@@ -9,6 +7,7 @@ import {
 } from "@navikt/aksel-icons";
 import { useTranslation } from "react-i18next";
 import classes from "./ToolBarItems.module.css";
+import { DraggableItem } from "../DraggableItem/DraggableItem";
 
 export const ToolBarItems = () => {
   const { t } = useTranslation();
@@ -49,40 +48,5 @@ export const ToolBarItems = () => {
         ),
       )}
     </>
-  );
-};
-
-export interface ItemProps {
-  icon: React.ReactNode;
-  text: string;
-  draggable: boolean;
-}
-
-interface DraggableItemProps {
-  item: ItemProps;
-  index: number;
-}
-
-const DraggableItem: React.FC<DraggableItemProps> = ({ item, index }) => {
-  const [{ isDragging }, drag] = useDrag(() => ({
-    type: DraggableItemsType.ToolbarItem,
-    item: { ...item, index },
-    collect: (monitor) => ({
-      isDragging: monitor.isDragging(),
-    }),
-  }));
-
-  return (
-    <div
-      ref={drag}
-      className={classes.toolbarIcon}
-      style={{
-        opacity: isDragging ? 0.5 : 1,
-        cursor: "move",
-      }}
-    >
-      {item.icon}
-      <div>{item.text}</div>
-    </div>
   );
 };
