+++ frontend/form-builder/src/components/Toolbar/Toolbar.tsx
@@ -1,5 +1,4 @@
 import { Heading } from "@digdir/design-system-react";
-
 import React from "react";
 import { useTranslation } from "react-i18next";
 import { ToolBarItems } from "./ToolBarItems/ToolBarItems";
@@ -8,17 +7,11 @@ export const Toolbar = (): React.JSX.Element => {
   const { t } = useTranslation();
 
   return (
-    <>
+    <div>
       <Heading level={2} size="medium" spacing>
         {t("toolbar_tools")}
       </Heading>
-      <ul>
-        <div>
-          <li>
-            <ToolBarItems />
-          </li>
-        </div>
-      </ul>
-    </>
+      <ToolBarItems />
+    </div>
   );
 };
